% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_skeleton.R
\name{calc_new_skeleton}
\alias{calc_new_skeleton}
\title{Function to update a CRM skeleton in a seamless design where the second
stage uses the CRM.}
\usage{
calc_new_skeleton(old_skeleton, tox, level, offset_seq = NULL)
}
\arguments{
\item{old_skeleton}{A numeric vector with entries between 0 and 1; the initial skeleton.}

\item{tox}{A vector with binary entries {0,1} representing the outcomes of the patients;
should be the same length as the number of observations collected.}

\item{level}{A positive integer vector of labels with length equal to that of tox indicating
the dose level that each observation was assigned to. Every element of this vector should,
in addition to being positive, be no larger than the length of skeleton, i.e. the number of
dose levels.}

\item{offset_seq}{A positive numeric vector of arbitrary length that should probably have
some values above and below 1. The old_skeleton corresponds to a value of 1, and including
offset_seq helps to see if a better fit is obtained by scaling old_skeleton up or down.}
}
\value{
A named list containing elements new_skeleton (the updated skeleton) and new_scale
(the suggested revised prior scale to put on the single parameter beta in the power model).
}
\description{
This function updates an initial CRM skeleton when using the empiric (power) model, after
some data have been gathered on the dose-toxicity curve. This code is currently not used,
but is included for posterity.
}
\references{
\insertRef{boonstra2020}{seamlesssim}

\insertRef{dfcrm2019}{seamlesssim}
}
