% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.titesim_ss.R
\name{print.titesim_ss}
\alias{print.titesim_ss}
\title{S3 print method for titesim_ss}
\usage{
\method{print}{titesim_ss}(x, ...)
}
\arguments{
\item{x}{An object of type titesim_ss, outputted from titesim_ss.}

\item{...}{List of additional arguments. Currently the following are implemented:}

\item{dgt}{Number of digits for output rounding. Default value is 3.}

\item{patient.detail}{Should patient-level information be provided, or only summary data?  Default is patient-level,
indicated by TRUE. For only summary data, use FALSE.}
}
\value{
Printed output from titesim_ss.
}
\description{
print.titesim_ss prints the results from titesim_ss, the simulator function
for a tite-crm trial.
}
\references{
\insertRef{boonstra2020}{seamlesssim}

\insertRef{dfcrm2019}{seamlesssim}
}
