% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamlesssim-package.R
\docType{package}
\name{seamlesssim-package}
\alias{seamlesssim-package}
\alias{seamlesssim}
\title{The 'seamlesssim' package.}
\description{
seamlesssim simulates complex seamless Phase I/II oncology trials
as discussed in the article by Boonstra et al. (2021). It allows clinical
trialists to determine operating characteristics of trials that assess both
toxicity and efficacy with a range of different design and analytic approaches.
For more information, see Boonstra et al. and the vignette.
}
\references{
\insertRef{boonstra2020}{seamlesssim}

\insertRef{boonstra2020b}{seamlesssim}

\insertRef{dfcrm2019}{seamlesssim}

\insertRef{rstan2020}{seamlesssim}
}
