% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_empiric_dec.R
\name{sim_empiric_dec}
\alias{sim_empiric_dec}
\title{A function to simulate a dose expansion cohort trial design}
\usage{
sim_empiric_dec(
  n_sim,
  true_tox_curve,
  stage_label = 1,
  sim_specific_start_id = NULL,
  sim_specific_dose_start = NULL,
  max_n_per_dec,
  module_rule = "local",
  thresh_decrease = 1/3,
  first_patient_look = 0,
  seed = sample(.Machine$integer.max, 1)
)
}
\arguments{
\item{n_sim}{How many simulated trials to conduct? (positive integer)}

\item{true_tox_curve}{A positive numeric vector that contains the true generating dose-toxicity
curve to simulate the data from. This also gives the number of dose levels.}

\item{stage_label}{A numeric value that can be arbitrary but is intended to take on integer
values equal to either 1 or 2, corresponding to the stage of the trial.}

\item{sim_specific_start_id}{A positive integer vector that contains the starting subject id
for each simulated trial. If provided, it must be as long as n_sim.}

\item{sim_specific_dose_start}{A positive integer vector that contains the starting dose level
for each simulated trial. If provided, it must be as long as n_sim and take on values between
0 (indicating that no patients should be enrolled) to length(true_tox_curve).}

\item{max_n_per_dec}{A positive integer giving the maximum enrollment for each DEC, if no
stopping for toxicity occurs.}

\item{module_rule}{Currently the only valid choice for this is the string "local", corresponding
to a local stopping rule: if the empiric proportion of DLTs at the current dose level ever
exceeds thresh_decrease, then de-escalate.}

\item{thresh_decrease}{A numeric value between 0 and 1. This is the maximum tolerance for the
observed proportion of DLTs at any given dose level. If this threshold is ever exceeded, the
DEC will de-escalate if possible or stop the trial entirely if not.}

\item{first_patient_look}{An integer greater than or equal to 0 and less than or equal to max_n_per_dec.
At a given dose level, at what point should the module_rule start taking action? This is an
ad-hoc way to prevent a scenario such as "The first patient experienced a DLT, therefore because
all patients at this dose level have experienced a DLT, we should de-escalate."}

\item{seed}{A positive integer seed for use prior to starting the simulations.}
}
\value{
A named list with entries:
\describe{
   \item{all_results}{A matrix giving the individual patient outcomes from all simulated trials.}
   \item{estMTD}{A vector as long as the number of simulated trials giving an integer value corresponding
   to the estimated MTD from each trial; a value of 0 indicates that all dose levels were estimated to be unsafe.}
   \item{enrollment}{A vector as long as the number of simulated trials giving the total enrollment for that trial.}
   \item{seed}{The seed that was used by the function.}
}
}
\description{
A general purpose simulator for dose expansion cohorts (DECs). It is intended to be called
from twostage_simulator rather than by the user directly.
}
\references{
\insertRef{boonstra2020}{seamlesssim}
}
