% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage_results.R
\name{twostage_results}
\alias{twostage_results}
\title{This function processes the output from twostage_simulator to produce several
basic plots and summary tables.}
\usage{
twostage_results(
  files = NULL,
  filepath = TRUE,
  primary_objectives = NULL,
  design_labels = NULL,
  scen_per_page = 10,
  design_per_page = 3
)
}
\arguments{
\item{files}{A character string giving the filepath to a folder containing
nothing but the raw output from twostage_simulator, saved as .Rds file(s). Alternatively, this can
a character vector containing the names of all of the raw output from twostage_simulator, stored
in the workspace.}

\item{filepath}{A logical value indicating whether files contains a filepath (filepath = TRUE) or
a character vector of file names stored in the workspace (filepath = FALSE). The default is TRUE.}

\item{primary_objectives}{A list containing three named elements: tox_target, tox_delta_no_exceed,
and eff_target, such that tox_target is between 0 and 1, tox_delta_no_exceed is between 0 and
(1 - tox_target), and eff_target is between 0 and 1. This should be the same list that was used in
twostage_simulator to generate the simulated trials.}

\item{design_labels}{A character vector giving labels for the different designs. If left NULL,
the number of the design will be used as its label. This should be the same length and the same order
as design_list inputted in twostage_simulator.}

\item{scen_per_page}{A numeric value indicating the number of data-generating scenarios that should
be printed per page. This can be at most 10 (the default is 10).}

\item{design_per_page}{A numeric value indicating the number of designs that should be printed per
page. This can be at most 3 (the default is 3).}
}
\value{
The function returns a named list containing two items: "plots" and "tables".
tables contains:
\describe{
   \item{generating_params_for_display}{This is a table giving the true efficacy and toxicity of
   each dose for each scenario considered.}
   \item{acc_dose_rec_table}{This is a table giving which proportion of trials for each design and
   scenario combination recommended an acceptable dose (a dose meeting the toxicity and efficacy
   standards, if not the best dose) at the end of the trial.}
   \item{mean_patients_table}{This is a table giving the mean number of patients enrolled for each
   design and scenario combination across the different simulated trials.}
}
plots contains:
\describe{
   \item{gen_param_plot}{This is the same information as in generating_params_for_display, but in
   plot form.}
   \item{acc_dose_rec_plot}{This is a plot giving which proportion of trials for each design and
   scenario combination recommended an acceptable dose, unacceptable dose, or made no recommendation
   at all.}
   \item{n_patients_plot}{This is a boxplot of the number of patients enrolled for each design and
   scenario combination across the different simulated trials.}
   \item{n_patients_RP2D_plot}{This is a boxplot of the number of patients who received the final
   recommended dose for each design and scenario combination across the different simulated trials.
   Note that the final recommended dose may not be safe or effective--this is merely a measure of
   how much patient data the design will yield for the final dose it recommends.}
   \item{prop_patients_acep_plot}{This is a boxplot of the number of patients who received an
   acceptable dose (a dose meeting the toxicity and efficacy standards, if not the best dose) for
   each design and scenario combination across the different simulated trials.}
}
}
\description{
For more information on function inputs and features, please see the vignette
and Boonstra (2020) on Arxiv.
}
