% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage_results.R
\name{twostage_results}
\alias{twostage_results}
\title{This function processes the output from twostage_simulator to produce several
basic plots and summary tables.}
\usage{
twostage_results(
  csv = FALSE,
  stage2folder = NULL,
  finaldatfolder = NULL,
  dose_outcome_curves = NULL,
  files = NULL,
  filepath = TRUE,
  primary_objectives = NULL,
  design_labels = NULL,
  scen_per_page = 10,
  design_per_page = 3
)
}
\arguments{
\item{csv}{A logical value indicating whether the output from twostage_simulator has been saved
as .csv file(s) (if .csv, then TRUE).}

\item{stage2folder}{A character string giving the filepath to a folder containing
nothing but sim_data_stage2 output from twostage_simulator, saved as .csv file(s).
This should only be provided if csv = TRUE.}

\item{finaldatfolder}{A character string giving the filepath to a folder containing
nothing but patient_data output from twostage_simulator, saved as .csv file(s).
This should only be provided if csv = TRUE.}

\item{dose_outcome_curves}{A list containing three named elements and an optional fourth
element: tox_curve, eff_curve, scenario, and, optionally, eff_curve_stage2. tox_curve is
the true toxicity curve of the doses; eff_curve is the true efficacy curve of the doses;
scenario is an identifier of which true data-generating scenario is being run (meant to be
helpful to the user when calling this function multiple times for different data-generating
scenarios). This should be the same list that was used in twostage_simulator to generate all
the simulated trials. This should only be provided if csv = TRUE.}

\item{files}{A character string giving the filepath to a folder containing
nothing but the raw output from twostage_simulator, saved as .Rds file(s). Alternatively, this can
a character vector containing the names of all of the raw output from twostage_simulator, stored
in the workspace. This should only be provided if csv = FALSE.}

\item{filepath}{A logical value indicating whether files contains a filepath (filepath = TRUE) or
a character vector of file names stored in the workspace (filepath = FALSE). The default is TRUE.
This should only be provided if csv = FALSE.}

\item{primary_objectives}{A list containing three named elements: tox_target, tox_delta_no_exceed,
and eff_target, such that tox_target is between 0 and 1, tox_delta_no_exceed is between 0 and
(1 - tox_target), and eff_target is between 0 and 1. This should be the same list that was used in
twostage_simulator to generate the simulated trials.}

\item{design_labels}{A character vector giving labels for the different designs. If left NULL,
the number of the design will be used as its label. This should be the same length and the same order
as design_list inputted in twostage_simulator.}

\item{scen_per_page}{A numeric value indicating the number of data-generating scenarios that should
be printed per page. This can be at most 10 (the default is 10).}

\item{design_per_page}{A numeric value indicating the number of designs that should be printed per
page. This can be at most 3 (the default is 3).}
}
\value{
The function returns a named list containing two items: "plots" and "tables".
tables contains:
\describe{
   \item{generating_params_for_display}{This is a table giving the true efficacy and toxicity of
   each dose for each scenario considered.}
   \item{acc_dose_rec_table}{This is a table giving which proportion of trials for each design and
   scenario combination recommended an acceptable dose (a dose meeting the toxicity and efficacy
   standards, if not the best dose) at the end of the trial.}
   \item{mean_patients_table}{This is a table giving the mean number of patients enrolled for each
   design and scenario combination across the different simulated trials.}
}
plots contains:
\describe{
   \item{gen_param_plot}{This is a list of plots providing the same information as in generating_params_for_display.
   If the number of designs under consideration exceeds design_per_page or the number of
   scenarios under consideration exceeds scen_per_page, the list will contain multiple plots. To call the
   first plot, we would run: results$plots$gen_param_plot[[1]] and so on for further plots.}
   \item{acc_dose_rec_plot}{This is a plot giving which proportion of trials for each design and
   scenario combination recommended an acceptable dose, unacceptable dose, or made no recommendation
   at all. If the number of designs under consideration exceeds design_per_page or the number of
   scenarios under consideration exceeds scen_per_page, the list will contain multiple plots. To call the
   first plot, we would run: results$plots$acc_dose_rec_param_plot[[1]] and so on for further plots.}
   \item{n_patients_plot}{This is a boxplot of the number of patients enrolled for each design and
   scenario combination across the different simulated trials.}
   \item{n_patients_RP2D_plot}{This is a boxplot of the number of patients who received the final
   recommended dose for each design and scenario combination across the different simulated trials.
   Note that the final recommended dose may not be safe or effective--this is merely a measure of
   how much patient data the design will yield for the final dose it recommends.}
   \item{prop_patients_accep_plot}{This is a boxplot of the number of patients who received an
   acceptable dose (a dose meeting the toxicity and efficacy standards, if not the best dose) for
   each design and scenario combination across the different simulated trials.}
}
}
\description{
For more information on function inputs and features, please see the vignette
and Boonstra (2020) on Arxiv.
}
